requirements.txt
web3
python-dotenv
firebase-admin
requests
google-cloud-firestore
